requirements.txt
appdirs==1.4.3
astropy==4.0
attrs==19.3.0
black==19.10b0
bumpversion==0.5.3
certifi==2019.11.28
cffi==1.13.2
Click==7.0
coverage==5.0.3
cryptography==2.8
dsnparse==0.1.15
entrypoints==0.3
Faker==3.0.0
filelock==3.0.12
flake8==3.7.9
importlib-metadata==1.4.0
mccabe==0.6.1
mock==3.0.5
more-itertools==8.1.0
mypy==0.761
mypy-extensions==0.4.3
mysql==0.0.2
mysqlclient==1.4.6
numpy==1.18.1
packaging==20.0
pandas==0.25.3
pathspec==0.7.0
Pillow==7.0.0
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyMySQL==0.9.3
pyparsing==2.4.6
pytest==5.3.2
pytest-mock==2.0.0
python-dateutil==2.8.1
pytz==2019.3
regex==2020.1.8
sentry-sdk==0.14.0
six==1.13.0
text-unidecode==1.3
toml==0.10.0
tox==3.14.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
virtualenv==16.7.9
wcwidth==0.1.8
zipp==1.0.0
